refactor: 요청한 메뉴 수와 주문된 메뉴 수가 다른 경우 발생하는 예외 로직 변경

+++ src/main/java/kitchenpos/domain/Order.java
@@ -46,8 +46,8 @@ public class Order {
         return new Order(id, orderTableId, OrderStatus.valueOf(orderStatus), orderedTime);
     }
 
-    public boolean isValidMenuSize(final int menuCount) {
-        return orderLineItems.size() == menuCount;
+    public boolean isValidMenuSize(final long menuCount) {
+        return (long) orderLineItems.size() == menuCount;
     }
 
     public void changeOrderStatus(final OrderStatus changeOrderStatus) {
